+++ Src/Notion.Client/Api/Blocks/BlocksRetrieveChildrenParameters.cs
@@ -1,0 +1,12 @@
+namespace Notion.Client
+{
+    public interface IBlocksRetrieveChildrenQueryParameters : IPaginationParameters
+    {
+    }
+
+    public class BlocksRetrieveChildrenParameters : IBlocksRetrieveChildrenQueryParameters
+    {
+        public string StartCursor { get; set; }
+        public string PageSize { get; set; }
+    }
+}
