+++ Test/Notion.UnitTests/BlocksClientTests.cs
@@ -24,7 +24,7 @@ namespace Notion.UnitTests
         {
             string blockId = "3c357473-a281-49a4-88c0-10d2b245a589";
 
-            var children = await _client.RetrieveChildrenAsync(blockId, new BlockRetrieveChildrenParameters());
+            var children = await _client.RetrieveChildrenAsync(blockId, new BlocksRetrieveChildrenParameters());
 
             Assert.NotNull(children);
         }
